rework, mileToKilometer

+++ source/mileToKilometer.cpp
@@ -1,15 +1,15 @@
 #include <iostream>
 #include <string>
 
-int calc(int mile){
-	int Kilometer=0;
+double calc(double mile){
+	double Kilometer=0;
 	Kilometer=mile/0.62137;
 	return Kilometer;
 	}
 
 int main()
 {
-	int mile;
+	double mile;
 	std::cout << "Bitte Meilen eingeben: " << std::endl;
 	std::cin >> mile;
 	std::cout << mile << " Meilen in Kilometer umgerechnet ergibt ca: " << calc(mile) << std::endl;
